flask_backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
mysql-connector-python
numpy
matplotlib
seaborn
PyJWT
werkzeug
python-dotenv
